requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
peewee==3.17.7
requests==2.32.3